% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_sum.R
\name{my_sum}
\alias{my_sum}
\title{Sum two vectors}
\usage{
my_sum(a, b)
}
\arguments{
\item{a}{An \code{integer()} or \code{numeric()} vector of length \code{L}.}

\item{b}{An \code{integer()} or \code{numeric()} vector of length \code{L}.}
}
\value{
An \code{integer()} or \code{numeric()} vector of length \code{L} with
the element-wise sum of \code{a} and \code{b}.
}
\description{
This function does the element wise sum of two vectors.
}
\details{
It really is just an example function that is powered by the \code{+} operator
from \link[base:Arithmetic]{base::Arithmetic}.
}
\examples{
## Generate some input data
x <- 1:10
y <- 1:10

## Perform the element wise sum
my_sum(x, y)
}
